move functions from MapSnapshotButton to mapActions
